+++ CHANGES.rst
@@ -96,6 +96,12 @@ New Features:
 
 New Features:
 
+- Convert all datetime, DateTime and time instances to UTC before serializing.
+  [thet]
+
+- Use python-dateutil instead of DateTime to parse date strings when de-serializing.
+  [thet]
+
 - Allow users to get their own user information.
   [erral]
 
@@ -110,7 +116,7 @@ Bugfixes:
 - Fix workflow translations with unicode characters.
   [Gagaro]
 
-- Fix workflow encoding in transition endpoint
+- Fix workflow encoding in transition endpoint.
   [Gagaro]